++ More Work on Character List.You can create and see : All your info exept Level,Race,Class, Hair number and Stuff.
git-svn-id: http://subversion.assembla.com/svn/sienna-rift@27 314f1d10-98f6-48e5-a4d7-766c1cb0bb34

+++ Shared/Remoting/CharacterMgr.cs
@@ -63,7 +63,7 @@ namespace Shared
             return CharacterDB.SelectObject<Character>("Name='" + CharacterDB.Escape(Name) + "' AND RealmId='" + RealmId + "'");
         }
 
-        public Character[] GetCharacters(int AccountId,byte RealmId)
+        public Character[] GetCharacters(long AccountId,byte RealmId)
         {
             return CharacterDB.SelectObjects<Character>("AccountId='" + AccountId + "' AND RealmId='" + RealmId + "'").ToArray();
         }
